Fix: dont error fully out when single api groupVersion not found (#68)
* Fix: dont error fully out when single api groupVersion not found

this commit changes the fail behavior when a particular groupVersion isnt found or is having issue (ie. service unavailable). Before we would error all the way back to the caller, now we will only warn that there was an issue. The return behavior remains the same, if we were unable to find the requested resource (or if there was an error calling the requested endpoint), the caller will get back undefined.

* Update kubeClass-test.js