Fix: bug where resources with same name dont both get kept (#114)
when multiple resources, due to the  creation of CRDs, have the same plural name, it results in an incomplete kubeResourceMeta list because of the way we stored the data. This change now will pull out crds into their own array in order to preserve all crds, but still allow for us to remove depricated "/apis/extensions" endpoints.